moved search bar logic to another file
